wip: misc changes to get GAR displayed in page
