+++ src/app/garfile/amg/index.js
@@ -2,6 +2,6 @@ const nav = require('../../../common/utils/nav');
 
 // Local dependencies
 const getController = require('./get.controller');
-const postController = require('./post.controller');
 
-module.exports = nav.buildRouterAndPaths('/garfile/review', getController, postController);
+
+module.exports = nav.buildRouterAndPaths('/garfile/amg', getController);
