+++ src/app/router.js
@@ -44,6 +44,7 @@ const garCancel = require('./garfile/cancel');
 const garsubmitsucess = require('./garfile/submit/success');
 const garsubmitfailure = require('./garfile/submit/failure');
 const printmanifest = require('./garfile/printmanifest');
+const amg  = require('./garfile/amg');
 
 // Saved entities
 const people = require('./people');
@@ -73,6 +74,9 @@ const editorganisation = require('./organisation/editorganisation');
 const orgassignrole = require('./organisation/assignrole');
 const orgEditUsers = require('./organisation/editusers');
 const deleteUser = require('./organisation/delete');
+
+// amg
+
 
 // Export
 module.exports.bind = (app) => {
@@ -135,4 +139,6 @@ module.exports.bind = (app) => {
   app.use(help.router);
   app.use(unavailable.router);
   app.use(printmanifest.router);
+  //
+  app.use(amg.router);
 };
